refactor: graphql-codegenのモックデータを別ファイルで生成
Fixes #456
